Change from Sec/Nsec StatInfo fields to time.Time().  The Stat_t
Timespec field names are not portable (even on unix systems, different
OSes use different field names).  Using the time.Time type is more
portable.

+++ results.go
@@ -168,7 +168,7 @@ func (r *Results) addMismatchedTotalBytes(size uint64) {
 	r.MismatchedTotalBytes += size
 }
 
-func (r *Results) foundInode(n uint32) {
+func (r *Results) foundInode(n uint64) {
 	r.InodeCount += 1
 	r.NlinkCount += int64(n)
 }
