+++ sortlinks_test.go
@@ -47,7 +47,7 @@ func setupInoStatInfo(fsdev *fsDev, inoSet I.Set) {
 		di, _ := I.LStatInfo(".")
 		// Deliberately make it so that if Nlinks are sorted, Inos are
 		// sorted also (for easier testing of []I.Ino result)
-		di.Nlink = uint32(ino)*2 + 100
+		di.Nlink = uint64(ino)*2 + 100
 		di.Ino = uint64(ino)
 		fsdev.inoStatInfo[I.Ino(ino)] = &di.StatInfo
 	}
